refactor(venom): spawn in a worker thread
Sometimes venom simply crashes while reading from the folder `tokens`.
For instance, putting venom in a new worker thread and waiting it ends
by timeout creates the possibility to delete the `tokens` folder and
to restart venom as necessary.

Several changes needed to be made to create this necessity, such as
changing the file hierarchy and fixing up the mess with the dockerfiles.

Still, the docker files does not have a way to create dependencies
between many images, however later on, it will be added two environments
for compose, namely developer and production.